Refactor into revealTransUnitTarget()
